+++ components.d.ts
@@ -6,7 +6,7 @@
  *
  * This is a generated file. Do not edit it's contents.
  *
- * This file was generated on 2023-02-27T05:00:10.823Z
+ * This file was generated on 2023-03-04T20:41:26.135Z
  */
 
 import { ChakraProps, chakra } from "@chakra-ui/vue-system"