+++ packages/c-reset/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @chakra-ui/c-reset
 
+## 2.1.0-beta.5
+
+### Minor Changes
+
+- bc95c4d6: Add popover component
+
+### Patch Changes
+
+- e4364213: Fix: Passthrough user-defined type attr for all button/iconbuttons
+- Updated dependencies [bc95c4d6]
+- Updated dependencies [e4364213]
+  - @chakra-ui/vue-system@2.1.0-beta.5
+
 ## 2.1.0-beta.4
 
 ### Patch Changes